requirements.txt
flask
dblpy
psutil
requests
wolframalpha
python-dotenv
discord.py>=1.3.4
requests_oauthlib
python-levenshtein
